requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
scikit-learn
numpy
pydantic
googletrans==4.0.0rc1
requests
python-dotenv